fix: Correctly report app version in built Docker images
A "COPY . /somewhere" instruction in a Dockerfile does not copy the
Dockerfile itself. This caused setuptools-scm inside the build container
to erroneously see a dirty worktree, leading it to generate a `X.Y.dev0`
version instead of the expected (and desired) `X.Y`.

This commit fixes the issue by restoring the Dockerfile from the git
index before starting the app build in the container.